+++ src/Beef.Core/Events/EventData.cs
@@ -14,6 +14,8 @@ namespace Beef.Events
     [JsonObject(MemberSerialization.OptIn)]
     public class EventData : IETag
     {
+        #region Obsolete
+
         /// <summary>
         /// Creates an <see cref="EventData"/> instance using a <see cref="Subject"/> <paramref name="template"/>.
         /// </summary>
@@ -21,6 +23,7 @@ namespace Beef.Events
         /// <param name="action">The event action.</param>
         /// <param name="keyValuePairs">The key/value pairs.</param>
         /// <returns>The <see cref="EventData"/>.</returns>
+        [Obsolete("The Template and KeyValuePair approach is being deprecated including all Create() methods; please use CreateEvent() or CreateValueEvent().")]
         public static EventData Create(string template, string action, params KeyValuePair<string, object?>[] keyValuePairs)
         {
             return ApplySubjectTemplate(new EventData(), Check.NotEmpty(template, nameof(template)), action, keyValuePairs);
@@ -35,6 +38,7 @@ namespace Beef.Events
         /// <param name="action">The event action.</param>
         /// <param name="keyValuePairs">The key/value pairs.</param>
         /// <returns>The <see cref="EventData{T}"/>.</returns>
+        [Obsolete("The Template and KeyValuePair approach is being deprecated including all Create() methods; please use CreateEvent() or CreateValueEvent().")]
         public static EventData<T> Create<T>(T value, string template, string action, params KeyValuePair<string, object?>[] keyValuePairs) where T : class
         {
             return (EventData<T>)ApplySubjectTemplate(new EventData<T> { Value = Check.NotNull(value, nameof(value)) }, Check.NotEmpty(template, nameof(template)), action, keyValuePairs);
@@ -43,9 +47,10 @@ namespace Beef.Events
         /// <summary>
         /// Applies the subject template.
         /// </summary>
+        [Obsolete("The Template and KeyValuePair approach is being deprecated including all Create() methods; please use CreateEvent() or CreateValueEvent().")]
         private static EventData ApplySubjectTemplate(EventData ed, string template, string action, params KeyValuePair<string, object?>[] keyValuePairs)
         {
-            ed.Subject = Event.CreateSubjectFromTemplate(template, keyValuePairs);
+            ed.Subject = PrependPrefix(Event.CreateSubjectFromTemplate(template, keyValuePairs));
             ed.Action = action;
 
             switch (keyValuePairs.Length)
@@ -71,12 +76,13 @@ namespace Beef.Events
         /// <param name="subject">The event subject.</param>
         /// <param name="action">The event action.</param>
         /// <returns>The <see cref="EventData"/>.</returns>
+        [Obsolete("The Template and KeyValuePair approach is being deprecated including all Create() methods; please use CreateEvent() or CreateValueEvent().")]
         public static EventData Create<T>(T value, string subject, string? action = null) where T : class
         {
             Check.NotNull(value, nameof(value));
 
             var ed = new EventData<T> { Value = value };
-            ed.Subject = Check.NotEmpty(subject, nameof(subject));
+            ed.Subject = PrependPrefix(Check.NotEmpty(subject, nameof(subject)));
             ed.Action = action;
 
             switch (value)
@@ -109,12 +115,91 @@ namespace Beef.Events
         /// <param name="subject">The event subject.</param>
         /// <param name="action">The event action.</param>
         /// <returns>The <see cref="EventData"/>.</returns>
+        [Obsolete("The Template and KeyValuePair approach is being deprecated including all Create() methods; please use CreateEvent() or CreateValueEvent().")]
         public static EventData Create(string subject, string? action = null)
         {
             var ed = new EventData { Action = action };
-            ed.Subject = Check.NotEmpty(subject, nameof(subject)); ;
+            ed.Subject = PrependPrefix(Check.NotEmpty(subject, nameof(subject)));
             return ed;
         }
+
+        #endregion
+
+        /// <summary>
+        /// Creates an <see cref="EventData"/> instance (with no <see cref="Key"/>.
+        /// </summary>
+        /// <param name="subject">The event subject.</param>
+        /// <param name="action">The event action.</param>
+        /// <returns>The <see cref="EventData"/>.</returns>
+        public static EventData CreateEvent(string subject, string? action = null) 
+            => new EventData { Subject = PrependPrefix(Check.NotEmpty(subject, nameof(subject))), Action = action };
+
+        /// <summary>
+        /// Creates an <see cref="EventData"/> instance with the specified <see cref="Key"/>.
+        /// </summary>
+        /// <param name="subject">The event subject.</param>
+        /// <param name="action">The event action.</param>
+        /// <param name="key">The event key.</param>
+        /// <returns>The <see cref="EventData"/>.</returns>
+        public static EventData CreateEvent(string subject, string? action = null, params IComparable?[] key) 
+            => new EventData { Subject = PrependPrefix(Check.NotEmpty(subject, nameof(subject))), Action = action, Key = key.Length == 1 ? (object?)key[0] : key };
+
+        /// <summary>
+        /// Creates an <see cref="EventData"/> instance using the <paramref name="value"/> (infers the <see cref="Key"/> from either <see cref="IIdentifier"/> or <see cref="IUniqueKey"/>).
+        /// </summary>
+        /// <typeparam name="T">The value <see cref="Type"/>.</typeparam>
+        /// <param name="value">The event value</param>
+        /// <param name="subject">The event subject.</param>
+        /// <param name="action">The event action.</param>
+        /// <returns>The <see cref="EventData"/>.</returns>
+        public static EventData<T> CreateValueEvent<T>(T value, string subject, string? action = null) where T : class
+        {
+            var ed = new EventData<T> { Value = Check.NotNull(value, nameof(value)) };
+            ed.Subject = PrependPrefix(Check.NotEmpty(subject, nameof(subject)));
+            ed.Action = action;
+
+            switch (value)
+            {
+                case IIntIdentifier ii:
+                    ed.Key = ii.Id;
+                    break;
+
+                case IGuidIdentifier gi:
+                    ed.Key = gi.Id;
+                    break;
+
+                case IStringIdentifier si:
+                    ed.Key = si.Id;
+                    break;
+
+                case IUniqueKey uk:
+                    if (uk.HasUniqueKey)
+                        ed.Key = uk.UniqueKey.Args.Length == 1 ? uk.UniqueKey.Args[0] : uk.UniqueKey.Args;
+                    else
+                        throw new InvalidOperationException("A Value that implements IUniqueKey must have one; i.e. HasUniqueKey = true.");
+
+                    break;
+            }
+
+            return ed;
+        }
+
+        /// <summary>
+        /// Creates an <see cref="EventData"/> instance with the specified <paramref name="value"/> <see cref="Key"/>.
+        /// </summary>
+        /// <typeparam name="T">The value <see cref="Type"/>.</typeparam>
+        /// <param name="value">The event value</param>
+        /// <param name="subject">The event subject.</param>
+        /// <param name="action">The event action.</param>
+        /// <param name="key">The event key.</param>
+        /// <returns>The <see cref="EventData"/>.</returns>
+        public static EventData<T> CreateValueEvent<T>(T value, string subject, string? action = null, params IComparable?[] key) 
+            => new EventData<T> { Value = value, Subject = PrependPrefix(Check.NotEmpty(subject, nameof(subject))), Action = action, Key = key.Length == 1 ? (object?)key[0] : key };
+
+        /// <summary>
+        /// Prepend the <see cref="Event.EventSubjectPrefix"/> to the <paramref name="subject"/> where specified.
+        /// </summary>
+        private static string PrependPrefix(string subject) => string.IsNullOrEmpty(Event.EventSubjectPrefix) ? subject : Event.EventSubjectPrefix + Event.PathSeparator + subject;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="EventData"/> class defaulting the <see cref="TenantId"/> and <see cref="Timestamp"/> as applicable