+++ src/Beef.Events/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 Represents the **NuGet** versions.
 
+## v3.1.2
+- *Enhancement:* Added `EventHubSubscriberHost.ExecutionContext(createFunc)` to more easily support the creation of a customised `ExecutionContext` instance.
+- *Enhancement:* The `IEventSubscriber.ReceiveAsync` must now (**breaking change**) return a `Result`; these include `Success()`, `DataNotFound()` (also automatically inferred from a `NotFoundException`), `InvalidData()` (also automatically inferred from a `ValidationException` or `BusinessException`). Otherwise, for anything else, just allow an `Exception` to bubble out.
+- *Enhancement:* `EventSubscriberHost` has been extended to support `NotSubscribedHandling` (defaults to `ContinueSilent`), `DataNotFoundHandling` (defaults to `Stop`) and `InvalidDataHandling` (defaults to `Stop`). `ResultHandling` options are `Stop`, `ContinueSilent`, `ContinueWithLogging` and `ContinueWithAudit`.
+
 ## v3.1.1
 - *Upgrade:* Upgraded the project to .NET Standard 2.1 (compatible with .NET Core 3.1).